+++ api/serializers/__init__.py
@@ -4,3 +4,4 @@ from .projects_field import ProjectsField
 from .new_threshold_field import NewThresholdField
 from .app_bookmark_field import AppBookmarkField
 from .tag_related_field import TagRelatedField
+from .identity_related_field import IdentityRelatedField
